finalCommit

+++ www/js/start/summaryCtrl.js
@@ -5,29 +5,41 @@
     .module('babyTracker')
     .controller('summaryCtrl', summaryCtrl);
 
-  summaryCtrl.$inject = ['Children', 'userDataSvc', 'childFollowerSvc', 'loggingService', '$scope', 'userSvc'];
+  summaryCtrl.$inject = ['Children', 'userDataSvc', 'childFollowerSvc', 'loggingService', '$scope', 'userSvc', '$ionicPopup'];
 
-  function summaryCtrl(Children, userDataSvc, childFollowerSvc, loggingService, $scope, userSvc) {
+  /**
+   * Controller that handles the child-summary view
+   */
+  function summaryCtrl(Children, userDataSvc, childFollowerSvc, loggingService, $scope, userSvc, $ionicPopup) {
     var vm = this;
+    //Variables
     vm.children = Children;
     vm.user = userDataSvc;
+
+    //Functions
     vm.doRefresh = doRefresh;
     vm.removeFollower = removeFollower;
-    activate();
-
-    function activate() {
-      console.log(Children);
-    }
 
     function removeFollower(child, index) {
-      console.log(child);
-      childFollowerSvc.deleteFollower(child.$id, userDataSvc.uid).then(function(res) {
-        var str = child.firstName + " " + child.lastName + " succesvol verwijderd";
-        loggingService.showSuccess(str, "user defollowed", "followerCtrl", false);
-        vm.children.splice(index, 1);
-      }, function(err) {
-        loggingService.showError("User removed failed", err, "followerCtrl", false);
-      })
+      var confirmPopup = $ionicPopup.confirm({
+        title: 'Verwijder kind',
+        template: 'Ben je zeker dat je ' + child.firstName + ' wil verwijderen?',
+        cancelText: 'Annuleer',
+        cancelType: 'button-positive',
+        okText: 'Verwijder',
+        okType: 'button-assertive'
+      });
+      confirmPopup.then(function(res) {
+        if (res) {
+          childFollowerSvc.deleteFollower(child.$id, userDataSvc.uid).then(function(res) {
+            var str = child.firstName + " " + child.lastName + " succesvol verwijderd";
+            loggingService.showSuccess(str, "user defollowed", "followerCtrl", true);
+            vm.children.splice(index, 1);
+          }, function(err) {
+            loggingService.showError("User removed failed", err, "followerCtrl", false);
+          });
+        }
+      });
     }
 
     function doRefresh() {
